requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
dataclasses==0.8
dnspython==2.1.0
Flask==2.0.1
idna==2.10
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pkg-resources==0.0.0
pycparser==2.20
pymongo==3.11.4
PyNaCl==1.4.0
requests==2.25.1
six==1.16.0
tinydb==4.4.0
typing-extensions==3.10.0.0
urllib3==1.26.7
Werkzeug==2.0.1
zipp==3.4.1
